Move Fabric code into Fabric directory
